+++ opentelemetry-sdk/src/opentelemetry/sdk/environment_variables.py
@@ -381,7 +381,7 @@ OTEL_EXPORTER_OTLP_METRICS_ENDPOINT = "OTEL_EXPORTER_OTLP_METRICS_ENDPOINT"
 """
 .. envvar:: OTEL_EXPORTER_OTLP_METRICS_ENDPOINT
 
-The :envvar:`OTEL_EXPORTER_OTLP_METRICS_ENDPOINT` target to which the span exporter is going to send spans.
+The :envvar:`OTEL_EXPORTER_OTLP_METRICS_ENDPOINT` target to which the metric exporter is going to send spans.
 The endpoint MUST be a valid URL host, and MAY contain a scheme (http or https), port and path.
 A scheme of https indicates a secure connection and takes precedence over this configuration setting.
 """
@@ -398,7 +398,7 @@ OTEL_EXPORTER_OTLP_METRICS_HEADERS = "OTEL_EXPORTER_OTLP_METRICS_HEADERS"
 """
 .. envvar:: OTEL_EXPORTER_OTLP_METRICS_HEADERS
 
-The :envvar:`OTEL_EXPORTER_OTLP_METRICS_HEADERS` contains the key-value pairs to be used as headers for spans
+The :envvar:`OTEL_EXPORTER_OTLP_METRICS_HEADERS` contains the key-value pairs to be used as headers for metrics
 associated with gRPC or HTTP requests.
 """
 
@@ -407,14 +407,14 @@ OTEL_EXPORTER_OTLP_METRICS_TIMEOUT = "OTEL_EXPORTER_OTLP_METRICS_TIMEOUT"
 .. envvar:: OTEL_EXPORTER_OTLP_METRICS_TIMEOUT
 
 The :envvar:`OTEL_EXPORTER_OTLP_METRICS_TIMEOUT` is the maximum time the OTLP exporter will
-wait for each batch export for spans.
+wait for each batch export for metrics.
 """
 
 OTEL_EXPORTER_OTLP_METRICS_COMPRESSION = "OTEL_EXPORTER_OTLP_METRICS_COMPRESSION"
 """
 .. envvar:: OTEL_EXPORTER_OTLP_METRICS_COMPRESSION
 
-Same as :envvar:`OTEL_EXPORTER_OTLP_COMPRESSION` but only for the span
+Same as :envvar:`OTEL_EXPORTER_OTLP_COMPRESSION` but only for the metric
 exporter. If both are present, this takes higher precendence.
 """